* feature/fix_release_issues:
  try with lower request rate
  add Rate limiting to Harvest API requests
  make tests more robust to existing data changing
  estimate item quantity is also a double
  throw RateLimitedException if we would have to wait more than 5 minutes to retry
  invoice item quantity is actually a double

+++ src/main/java/ch/aaap/harvestclient/domain/EstimateItem.java
@@ -41,7 +41,7 @@ public interface EstimateItem extends Reference<EstimateItem> {
     String getDescription();
 
     @Nullable
-    Long getQuantity();
+    Double getQuantity();
 
     Double getUnitPrice();
