+++ src/main/java/ch/aaap/harvestclient/domain/param/InvoiceItemUpdateInfo.java
@@ -28,7 +28,7 @@ public interface InvoiceItemUpdateInfo {
     String getDescription();
 
     @Nullable
-    Long getQuantity();
+    Double getQuantity();
 
     @Nullable
     Double getUnitPrice();
